Install script finished
file locking fixes for logrecorder
radioreference API update information logic finished
moved calls database to its own file to prevent locking
frameworked install script
added unitrunker XML importing

+++ ScanEyesV3/index.php
@@ -1,12 +1,21 @@
 <?php
 // General includes
 include 'includes/session.php'; // load session activator and reputation set
+if (is_dir('install') && is_file('../database/config.sqlite') == FALSE) {
+	echo "<a href='./install/'>Please configure ScanEyes first</a>";
+	exit();
+}
 include 'includes/classes.php'; // load database class
 include 'includes/config.php'; // read variables from database
 include 'includes/head.php'; //<head> tag
 
 // Page specific libraries
 include 'libraries/gen-sec.php';
+include 'libraries/gen-gen.php';
+
+if (is_dir('install') && is_file('../database/config.sqlite') == TRUE) {
+	rrmdir('install'); // If database exists, but install dir also exists, remove install dir
+}
 
 echo "usrlvl: ".$_SESSION['usrlvl']." rep:".$_SESSION['reputation']."<br>"; //debug
 if (@!$_GET['page']) { // If a user just types in the website name without page name
@@ -36,8 +45,20 @@ if ($page == "home") {
 }elseif ($page == "login") {
 	include 'gen/login.php';
 
-}elseif ($page == "importrr") {
-	include 'gen/importrr.php';
+}elseif ($page == "importrrtgid") {
+	include 'admin/backend/importrrtgid.php';
+
+}elseif ($page == "rrupdatetgid") {
+	include 'admin/backend/rrupdatetgid.php';
+
+}elseif ($page == "importuttgid") {
+	include 'admin/backend/importuttgid.php';
+
+}elseif ($page == "utupdatetgid") {
+	include 'admin/backend/utupdatetgid.php';
+
+}elseif ($page == "viewsystem") {
+	include 'gen/viewsystem.php';
 
 }elseif ($page == "logoff") {
 	include 'gen/logoff.php';
